requirements-deploy.txt
# Core dependencies
python-dotenv==1.0.0
aiohttp==3.12.14
lxml==6.0.0
beautifulsoup4==4.13.4
requests==2.32.4

# Database - MySQL only
mysql-connector-python==8.2.0

# Telegram bot (using older version for compatibility)
python-telegram-bot==13.15

# Web scraping
selenium==4.34.0
selenium-stealth==1.0.6
pyperclip==1.9.0

# Async support
trio==0.30.0
httpx==0.28.1

# Other utilities
websocket-client==1.8.0 